requirement.txt
streamlit
psycopg2-binary
SQLAlchemy
pandas
fpdf2